Python_appka/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
Markdown==3.4.3
Pygments==2.14.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
